mcp-servers/document-accessibility-server/requirements.txt
mcp==1.1.2
pypdf==5.1.0
pdfplumber==0.11.4
PyMuPDF==1.24.14
python-docx==1.1.2
Pillow==11.0.0
